esconder/requirements.txt
pip==23.2.1
setuptools==68.0.0
blinker==1.6.2
click==8.1.7
colorama==0.4.6
Flask==2.3.3
gunicorn==22.0.0
itsdangerous==2.1.2
Jinja2==3.1.4
MarkupSafe==2.1.3
packaging==23.1
python-decouple==3.8
Werkzeug==2.3.7
python-dotenv==1.0.0
requests==2.32.2
hcaptcha==0.2.1
aiohttp>=3.9.4 # not directly required, pinned by Snyk to avoid a vulnerability
urllib3>=2.2.2 # not directly required, pinned by Snyk to avoid a vulnerability
